requirements.txt
streamlit
anthropic
pytrends
requests
python-dotenv
pandas
